apps/backend/requirements.txt
# Core FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
redis==5.2.1
aiohttp==3.11.10
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.21.0

# Authentication & Security
argon2-cffi==25.1.0
authlib==1.6.4
itsdangerous==2.2.0
PyJWT==2.10.1
bleach==6.2.0

# Email
email-validator==2.3.0
aiosmtplib==4.0.2
jinja2==3.1.6

# AI Providers
openai==1.57.2
httpx==0.28.1
certifi==2024.12.14
backoff==2.2.1

# WebSockets & SSE
websockets==15.0.1
sse-starlette==2.2.1

# Background Tasks & CLI
celery==5.5.3
click==8.3.0
schedule==1.2.2

# Image Processing (Optional)
Pillow==11.3.0

# Cloud Storage (Optional)
boto3==1.40.40

# Monitoring & Production
prometheus_client==0.21.1
docker==7.1.0
pyyaml==6.0.2
aiofiles==24.1.0

# Development dependencies
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
mypy==1.18.2
ruff==0.13.2
black==25.9.0
psutil==7.1.0

# Error Tracking
sentry-sdk[fastapi]==2.20.0
